Add unsigned option to /transaction/verify
