+++ src/api/integration/testdata/verify-transaction-invalid-empty.golden
@@ -1,4 +1,5 @@
 {
+	"unsigned": false,
 	"confirmed": false,
 	"transaction": {
 		"length": 0,
