+++ src/daemon/gateway.go
@@ -634,6 +634,6 @@ func (gw *Gateway) GetHealth() (*Health, error) {
 
 // VerifyTxnVerbose verifies an isolated transaction and returns []wallet.UxBalance of
 // transaction inputs, whether the transaction is confirmed and error if any
-func (gw *Gateway) VerifyTxnVerbose(txn *coin.Transaction) ([]visor.TransactionInput, bool, error) {
-	return gw.v.VerifyTxnVerbose(txn)
+func (gw *Gateway) VerifyTxnVerbose(txn *coin.Transaction, signed visor.TxnSignedFlag) ([]visor.TransactionInput, bool, error) {
+	return gw.v.VerifyTxnVerbose(txn, signed)
 }
